Added ability to filter Vagrant's output for data
In the output from `vagrant up` is information that could be critical to
programatically interfacing with the machine once it's running. This
commit adds a new parameter to `Vagrant.up()` called `output_filter`,
which allows the user to specify a set of regular expressions to look
for in the command's output, and which is returned as a dict. The
user-specified output context manager (given when creating the Vagrant
instance) will still continue to receive the output of the command as
well.